[UPDATE] Initial work for blockchain implementation completed
