+++ src/test_blockchain.py
@@ -1,3 +1,11 @@
 #!/usr/bin/python3
 
 # Author: @andrewk10
+
+# Importing blockchain for blockchain  based functionality
+import blockchain
+
+
+def_test_create_blockchain:
+    test_blockchain = blockchain
+
